splitting watcher functionality into a separate file
